requirements.txt
click>=8.0.0
fpdf2>=2.7.5
matplotlib>=3.5.0
opencv-python>=4.5.5.64
numpy>=1.21.0
psutil>=5.9.0
Pillow>=9.0.0

# ML and FastVLM
torch>=2.0.0
transformers>=4.30.0
accelerate>=0.20.0
sentencepiece>=0.1.99
protobuf>=3.20.0

# Video processing
ffmpeg-python>=0.2.0

# Utilities
tqdm>=4.65.0
python-dotenv>=0.19.0
